+++ src/components/VvDialog/VvDialog.vue
@@ -9,9 +9,10 @@
 	import VvIcon from '../VvIcon/VvIcon.vue'
 	import { VvDialogEvents, VvDialogProps } from '.'
 
-	// props and emit
+	// props, emit and template refs
 	const props = defineProps(VvDialogProps)
 	const emit = defineEmits(VvDialogEvents)
+	const dialogEl: Ref<HTMLDialogElement | undefined> = ref()
 
 	// data
 	const localModelValue = ref(false)
@@ -26,7 +27,6 @@
 			emit('update:modelValue', value)
 		},
 	})
-	const htmlAttrIsOpen = ref(true)
 
 	// template ref
 	const modalWrapper = ref(null)
@@ -36,7 +36,6 @@
 		const { id } = props
 		return {
 			id,
-			open: htmlAttrIsOpen.value,
 		} as DialogHTMLAttributes
 	})
 	const dialogClass = computed(() => {
@@ -50,11 +49,11 @@
 	const transitioName = computed(() => `vv-dialog--${props.transition}`)
 	const dialogTransitionHandlers = {
 		'before-enter': () => {
-			htmlAttrIsOpen.value = true
+			dialogEl.value?.showModal()
 			emit('open')
 		},
 		'after-leave': () => {
-			htmlAttrIsOpen.value = false
+			dialogEl.value?.close()
 			emit('close')
 		},
 	}
@@ -87,7 +86,12 @@
 
 <template>
 	<Transition :name="transitioName" v-on="dialogTransitionHandlers">
-		<dialog v-show="modelValue" v-bind="dialogAttrs" :class="dialogClass">
+		<dialog
+			v-show="modelValue"
+			v-bind="dialogAttrs"
+			ref="dialogEl"
+			:class="dialogClass"
+		>
 			<article ref="modalWrapper" class="vv-dialog__wrapper">
 				<header v-if="$slots.header || title" class="vv-dialog__header">
 					<!-- @slot Header slot -->